requirements.txt
streamlit
google-generativeai
python-dotenv
poppler-utils
PyPDF2
streamlit-lottie
requests
fpdf
matplotlib
wordcloud

